Sync docs, further refinements (#4263)
More improvements to sync docs before announcement.

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [X] `other`

### Release notes

---------
