+++ packages/sync-core/src/lib/ClientWebSocketAdapter.ts
@@ -1,6 +1,6 @@
 import { atom, Atom } from '@tldraw/state'
 import { TLRecord } from '@tldraw/tlschema'
-import { assert } from '@tldraw/utils'
+import { assert, warnOnce } from '@tldraw/utils'
 import { chunk } from './chunk'
 import { TLSocketClientSentEvent, TLSocketServerSentEvent } from './protocol'
 import {
@@ -71,7 +71,11 @@ export class ClientWebSocketAdapter implements TLPersistentClientSocket<TLRecord
 		this._reconnectManager.connected()
 	}
 
-	private _handleDisconnect(reason: 'closed' | 'error' | 'manual', closeCode?: number) {
+	private _handleDisconnect(
+		reason: 'closed' | 'error' | 'manual',
+		closeCode?: number,
+		didOpen?: boolean
+	) {
 		debug('handleDisconnect', {
 			currentStatus: this.connectionStatus,
 			closeCode,
@@ -95,6 +99,12 @@ export class ClientWebSocketAdapter implements TLPersistentClientSocket<TLRecord
 				break
 		}
 
+		if (closeCode === 1006 && !didOpen) {
+			warnOnce(
+				"Could not open WebSocket connection. This might be because you're trying to load a URL that doesn't support websockets. Check the URL you're trying to connect to."
+			)
+		}
+
 		if (
 			// it the status changed
 			this.connectionStatus !== newStatus &&
@@ -116,6 +126,9 @@ export class ClientWebSocketAdapter implements TLPersistentClientSocket<TLRecord
 				this._ws.readyState === WebSocket.CLOSING,
 			`Tried to set a new websocket in when the existing one was ${this._ws?.readyState}`
 		)
+
+		let didOpen = false
+
 		// NOTE: Sockets can stay for quite a while in the CLOSING state. This is because the transition
 		//       between CLOSING and CLOSED happens either after the closing handshake, or after a
 		//       timeout, but in either case those sockets don't need any special handling, the browser
@@ -126,18 +139,19 @@ export class ClientWebSocketAdapter implements TLPersistentClientSocket<TLRecord
 				this._ws === ws,
 				"sockets must only be orphaned when they are CLOSING or CLOSED, so they can't open"
 			)
+			didOpen = true
 			this._handleConnect()
 		}
 		ws.onclose = (event: CloseEvent) => {
-			debug('ws.onclose')
+			debug('ws.onclose', event)
 			if (this._ws === ws) {
-				this._handleDisconnect('closed', event.code)
+				this._handleDisconnect('closed', event.code, didOpen)
 			} else {
 				debug('ignoring onclose for an orphaned socket')
 			}
 		}
-		ws.onerror = () => {
-			debug('ws.onerror')
+		ws.onerror = (event) => {
+			debug('ws.onerror', event)
 			if (this._ws === ws) {
 				this._handleDisconnect('error')
 			} else {
