+++ templates/sync-cloudflare/client/App.tsx
@@ -1,13 +1,7 @@
 import { useSync } from '@tldraw/sync'
-import {
-	AssetRecordType,
-	getHashForString,
-	TLAsset,
-	TLAssetStore,
-	TLBookmarkAsset,
-	Tldraw,
-	uniqueId,
-} from 'tldraw'
+import { Tldraw } from 'tldraw'
+import { getBookmarkPreview } from './getBookmarkPreview'
+import { multiplayerAssetStore } from './multiplayerAssetStore'
 
 // Where is our worker located? Configure this in `vite.config.ts`
 const WORKER_URL = process.env.TLDRAW_WORKER_URL
@@ -21,7 +15,7 @@ function App() {
 		// We need to know the websockets URI...
 		uri: `${WORKER_URL}/connect/${roomId}`,
 		// ...and how to handle static assets like images & videos
-		assets: multiplayerAssets,
+		assets: multiplayerAssetStore,
 	})
 
 	return (
@@ -32,69 +26,11 @@ function App() {
 				store={store}
 				onMount={(editor) => {
 					// when the editor is ready, we need to register our bookmark unfurling service
-					editor.registerExternalAssetHandler('url', unfurlBookmarkUrl)
+					editor.registerExternalAssetHandler('url', getBookmarkPreview)
 				}}
 			/>
 		</div>
 	)
-}
-
-// How does our server handle assets like images and videos?
-const multiplayerAssets: TLAssetStore = {
-	// to upload an asset, we prefix it with a unique id, POST it to our worker, and return the URL
-	async upload(_asset, file) {
-		const id = uniqueId()
-
-		const objectName = `${id}-${file.name}`.replace(/[^a-zA-Z0-9.]/g, '-')
-		const url = `${WORKER_URL}/uploads/${objectName}`
-
-		const response = await fetch(url, {
-			method: 'POST',
-			body: file,
-		})
-
-		if (!response.ok) {
-			throw new Error(`Failed to upload asset: ${response.statusText}`)
-		}
-
-		return url
-	},
-	// to retrieve an asset, we can just use the same URL. you could customize this to add extra
-	// auth, or to serve optimized versions / sizes of the asset.
-	resolve(asset) {
-		return asset.props.src
-	},
-}
-
-// How does our server handle bookmark unfurling?
-async function unfurlBookmarkUrl({ url }: { url: string }): Promise<TLAsset> {
-	const asset: TLBookmarkAsset = {
-		id: AssetRecordType.createId(getHashForString(url)),
-		typeName: 'asset',
-		type: 'bookmark',
-		meta: {},
-		props: {
-			src: url,
-			description: '',
-			image: '',
-			favicon: '',
-			title: '',
-		},
-	}
-
-	try {
-		const response = await fetch(`${WORKER_URL}/unfurl?url=${encodeURIComponent(url)}`)
-		const data = await response.json()
-
-		asset.props.description = data?.description ?? ''
-		asset.props.image = data?.image ?? ''
-		asset.props.favicon = data?.favicon ?? ''
-		asset.props.title = data?.title ?? ''
-	} catch (e) {
-		console.error(e)
-	}
-
-	return asset
 }
 
 export default App
